ffill and bfill for intervals
