+++ tests/series/test_series.py
@@ -1538,23 +1538,6 @@ def test_types_ne() -> None:
     check(assert_type(s1 != s2, "pd.Series[bool]"), pd.Series, np.bool)
 
 
-def test_types_bfill() -> None:
-    s1 = pd.Series([1, 2, 3])
-    check(assert_type(s1.bfill(), "pd.Series[int]"), pd.Series, np.integer)
-    check(assert_type(s1.bfill(inplace=False), "pd.Series[int]"), pd.Series, np.integer)
-    check(
-        assert_type(s1.bfill(inplace=False, limit_area="inside"), "pd.Series[int]"),
-        pd.Series,
-        np.integer,
-    )
-    check(assert_type(s1.bfill(inplace=True), "pd.Series[int]"), pd.Series, np.integer)
-    check(
-        assert_type(s1.bfill(inplace=True, limit_area="outside"), "pd.Series[int]"),
-        pd.Series,
-        np.integer,
-    )
-
-
 def test_types_ewm() -> None:
     s1 = pd.Series([1, 2, 3])
     check(
@@ -1589,23 +1572,6 @@ def test_types_ewm() -> None:
             "ExponentialMovingWindow[pd.Series]",
         ),
         ExponentialMovingWindow,
-    )
-
-
-def test_types_ffill() -> None:
-    s1 = pd.Series([1, 2, 3])
-    check(assert_type(s1.ffill(), "pd.Series[int]"), pd.Series, np.integer)
-    check(assert_type(s1.ffill(inplace=False), "pd.Series[int]"), pd.Series, np.integer)
-    check(
-        assert_type(s1.ffill(inplace=False, limit_area="inside"), "pd.Series[int]"),
-        pd.Series,
-        np.integer,
-    )
-    check(assert_type(s1.ffill(inplace=True), "pd.Series[int]"), pd.Series, np.integer)
-    check(
-        assert_type(s1.ffill(inplace=True, limit_area="outside"), "pd.Series[int]"),
-        pd.Series,
-        np.integer,
     )
